Call reset on new plan, not new goal
The `reset()` functions were called in `setGoalPose()`, i.e. when the
user gives a new goal. However, when the local planner fails and the
global planner replans, `setGoalPose()` is not called, but `setPlan()`
is. This commit makes sure that `reset()` is also called in this case.

This is important in situations where e.g. the robot is already in the
goal pose, but cannot turn to the final orientation because of
obstacles. The new global plan might require the robot to leave the goal
pose again, and so the critics (like for example the RotateToGoal
critic) need to be reset.